Extract and expose Splash component for visualizations website
