+++ src/Splash/README.md
@@ -1,0 +1,16 @@
+The splash component can be used to render a simple animated splash screen for a project. It is used internally by this package and by https://github.com/contiamo/operational-visualizations.
+
+### Usage
+
+```jsx static
+<Splash
+  title="Some Title"
+  actions={
+    <>
+      <Button>Action</Button>
+    </>
+  }
+>
+  <p>A project</p>
+</Splash>
+```
