+++ src/Splash/Splash.Logo.tsx
