+++ src/index.ts
@@ -1,38 +1,39 @@
 import OperationalUI from "./OperationalUI/OperationalUI"
 
 export { OperationalStyleConstants } from "./utils/constants"
+
 export { default as ActionMenu, ActionMenuProps } from "./ActionMenu/ActionMenu"
 export { default as Avatar, AvatarProps } from "./Avatar/Avatar"
 export { default as AvatarGroup, AvatarGroupProps } from "./AvatarGroup/AvatarGroup"
 export { default as Breadcrumb, BreadcrumbProps } from "./Breadcrumb/Breadcrumb"
 export { default as Breadcrumbs, BreadcrumbsProps } from "./Breadcrumbs/Breadcrumbs"
+export { default as Breakdown, BreakdownProps } from "./Breakdown/Breakdown"
 export { default as Button, ButtonProps } from "./Button/Button"
 export { default as ButtonGroup, ButtonGroupProps } from "./ButtonGroup/ButtonGroup"
-export { default as Breakdown, BreakdownProps } from "./Breakdown/Breakdown"
 export { default as Card, CardProps } from "./Card/Card"
-export { default as CardItem, CardItemProps } from "./CardItem/CardItem"
-export { default as CardHeader, CardHeaderProps } from "./CardHeader/CardHeader"
 export { default as CardColumn, CardColumnProps } from "./CardColumn/CardColumn"
 export { default as CardColumns } from "./CardColumns/CardColumns"
-export { default as InfoPanel, InfoPanelProps } from "./InfoPanel/InfoPanel"
+export { default as CardHeader, CardHeaderProps } from "./CardHeader/CardHeader"
+export { default as CardItem, CardItemProps } from "./CardItem/CardItem"
 export { default as Chip, ChipProps } from "./Chip/Chip"
 export { default as Code, CodeProps } from "./Code/Code"
 export { default as ContextMenu, ContextMenuProps } from "./ContextMenu/ContextMenu"
-export { default as Logo, LogoProps } from "./Logo/Logo"
-export { default as Grid, GridProps } from "./Grid/Grid"
-export { default as HeaderMenu, HeaderMenuProps } from "./HeaderMenu/HeaderMenu"
-export { default as Hint, HintProps } from "./Hint/Hint"
 export { default as DatePicker, DatePickerProps } from "./DatePicker/DatePicker"
 export { default as Form } from "./Form/Form"
+export { default as Grid, GridProps } from "./Grid/Grid"
 export { default as HeaderBar, HeaderBarProps } from "./HeaderBar/HeaderBar"
-export { default as Layout, LayoutProps } from "./Layout/Layout"
+export { default as HeaderMenu, HeaderMenuProps } from "./HeaderMenu/HeaderMenu"
+export { default as Hint, HintProps } from "./Hint/Hint"
 export { default as Icon, IconName, IconProps } from "./Icon/Icon"
+export { default as InfoPanel, InfoPanelProps } from "./InfoPanel/InfoPanel"
 export { default as Input, InputProps } from "./Input/Input"
+export { default as Layout, LayoutProps } from "./Layout/Layout"
+export { default as Logo, LogoProps } from "./Logo/Logo"
 export { default as Message, MessageProps } from "./Message/Message"
 export { default as Messages, MessagesProps } from "./Messages/Messages"
 export { default as NameTag, NameTagProps } from "./NameTag/NameTag"
-export { default as OperationalUI, OperationalUIProps } from "./OperationalUI/OperationalUI"
 export { default as OperationalContext, Context } from "./OperationalContext/OperationalContext"
+export { default as OperationalUI, OperationalUIProps } from "./OperationalUI/OperationalUI"
 export { default as Page, PageProps } from "./Page/Page"
 export { default as PageArea, PageAreaProps } from "./PageArea/PageArea"
 export { default as PageAreas, PageAreasProps } from "./PageAreas/PageAreas"
@@ -43,19 +44,21 @@ export { default as ProgressPanel, ProgressPanelProps } from "./ProgressPanel/Pr
 export { default as ResourceName } from "./ResourceName/ResourceName"
 export { default as Select, SelectProps } from "./Select/Select"
 export { default as Sha } from "./Sha/Sha"
-export { default as FinePrint } from "./Typography/FinePrint"
-export { default as Body } from "./Typography/Body"
-export { default as Title } from "./Typography/Title"
-export { default as Small } from "./Typography/Small"
 export { default as Sidenav, SidenavProps } from "./Sidenav/Sidenav"
 export { default as SidenavHeader, SidenavHeaderProps } from "./SidenavHeader/SidenavHeader"
 export { default as SidenavItem, SidenavItemProps } from "./SidenavItem/SidenavItem"
 export { default as Spinner, SpinnerProps } from "./Spinner/Spinner"
+export { default as Splash, SplashProps } from "./Splash/Splash"
 export { default as Status, StatusProps } from "./Status/Status"
 export { default as Switch, SwitchProps } from "./Switch/Switch"
 export { default as Table, TableProps } from "./Table/Table"
 export { default as Textarea, TextareaProps } from "./Textarea/Textarea"
 export { default as Tooltip, TooltipProps } from "./Tooltip/Tooltip"
 export { default as Tree, TreeProps } from "./Tree/Tree"
+
+export { default as FinePrint } from "./Typography/FinePrint"
+export { default as Body } from "./Typography/Body"
+export { default as Title } from "./Typography/Title"
+export { default as Small } from "./Typography/Small"
 
 export default OperationalUI
