--- src/website/index.ts
@@ -1,3 +1,0 @@
-import Splash from "./Splash"
-
-export default Splash
